[Feat]#460 진행률 기록 API 구현

+++ src/main/java/com/example/bookiibookii/domain/memberbook/dto/res/LibraryMemberBookResponseDTO.java
@@ -27,7 +27,7 @@ public class LibraryMemberBookResponseDTO {
     private LocalDate startDate;
     private LocalDate endDate;
     private Integer duration;
-    private Double progressRate;
+    private Integer progressRate;
     private Double rating;
     private String comment;
 }
